Implement more `__*printf_chk` functions.
